+++ packages/editor/editor-core/src/__tests__/unit/plugins/status/nodeviews/status.tsx
@@ -9,7 +9,7 @@ import {
   dispatchPasteEvent,
   status,
 } from '@atlaskit/editor-test-helpers';
-import { Status } from '@atlaskit/status';
+import { Status } from '@atlaskit/status/element';
 import StatusNodeView, {
   Props as StatusNodeViewProps,
   State as StatusNodeViewState,