+++ packages/editor/editor-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-core
 
+## 107.10.0
+- [minor] [969915d261](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/969915d261):
+
+  - update status import entrypoints in Renderer/editor
+
 ## 107.9.6
 - [patch] [0ff405bd0f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0ff405bd0f):