+++ packages/editor/renderer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/renderer
 
+## 45.4.0
+- [minor] [969915d261](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/969915d261):
+
+  - update status import entrypoints in Renderer/editor
+
 ## 45.3.3
 - [patch] [32317ff8f3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/32317ff8f3):